requirements.txt
amqp==5.0.6
appnope==0.1.2
arabic-reshaper==2.1.3
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.4.0
black==21.11b1
boto3==1.17.105
botocore==1.20.105
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.3
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.44.0
django-celery-beat==2.2.1
django-countries==7.2.1
django-debug-toolbar==3.2.1
django-decouple==2.1
django-filter==2.4.0
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.8
ipython==7.25.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.18.0
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.4
matplotlib-inline==0.1.2
mypy-extensions==0.4.3
oauthlib==3.1.1
openpyxl==3.0.7
parso==0.8.2
pathspec==0.9.0
paystackapi==2.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.0
platformdirs==2.4.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
regex==2021.11.10
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
sentry-sdk==1.1.0
six==1.16.0
sqlparse==0.4.1
suds-community==1.0.0
tomli==1.2.2
traitlets==5.0.5
typing_extensions==4.0.0
untangle==1.1.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xmltodict==0.12.0
